utils/requirements.txt
psycopg2_binary==2.9.5